requirements.txt
a2wsgi==1.10.10
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==3.0.6
apache-airflow-core==3.0.6
apache-airflow-providers-common-compat==1.7.3
apache-airflow-providers-common-io==1.6.2
apache-airflow-providers-common-sql==1.27.5
apache-airflow-providers-smtp==2.2.0
apache-airflow-providers-standard==1.6.0
apache-airflow-task-sdk==1.0.6
appnope==0.1.4
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
azure-functions==1.23.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
build==1.2.1
CacheControl==0.14.0
cadwyn==5.4.4
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.2.1
colorlog==6.9.0
comm==0.2.2
crashtest==0.4.1
cron-converter==1.2.2
cron_descriptor==2.0.5
croniter==6.0.0
cryptography==45.0.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distlib==0.3.8
dnspython==2.7.0
dulwich==0.21.7
email-validator==2.3.0
executing==2.1.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.1.2
fqdn==1.5.1
fsspec==2025.9.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.6
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyring==24.3.1
lazy-object-proxy==1.12.0
libcst==1.8.2
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
methodtools==0.4.7
mistune==3.0.2
more-itertools==10.4.0
msgpack==1.0.8
msgspec==0.19.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.3.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
overrides==7.7.0
packaging==24.1
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
pipenv==2024.4.0
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.5
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.10.1
pyproject_hooks==1.1.0
pyspark==3.3.4
pytest==8.3.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.7
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rich-argparse==1.7.1
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.20.0
Send2Trash==1.8.3
sentry-sdk==2.35.2
setproctitle==1.3.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
trove-classifiers==2024.7.2
typer==0.17.3
types-python-dateutil==2.9.0.20240821
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uri-template==1.3.0
urllib3==2.2.2
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.26.3
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wirerope==1.0.0
wrapt==1.17.3
xattr==1.1.0
zipp==3.20.1
